subiquity_client: hide logger object instances
Don't leak subiquity_client's logger instances to prevent accidental
use of a wrong logger instance in the installer.
